:recycle: refactor move competence DS to `src`
