+++ api/lib/infrastructure/datasources/learning-content/index.js
@@ -1,6 +1,6 @@
 import { areaDatasource } from './area-datasource.js';
 import { challengeDatasource } from './challenge-datasource.js';
-import { competenceDatasource } from './competence-datasource.js';
+import { competenceDatasource } from '../../../../src/shared/infrastructure/datasources/learning-content//competence-datasource.js';
 import { courseDatasource } from './course-datasource.js';
 import { frameworkDatasource } from './framework-datasource.js';
 import { skillDatasource } from './skill-datasource.js';
