+++ api/src/shared/infrastructure/repositories/competence-repository.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import { LearningContentResourceNotFound } from '../datasources/learning-content/LearningContentResourceNotFound.js';
 import { Competence } from '../../domain/models/Competence.js';
-import { competenceDatasource } from '../../../../lib/infrastructure/datasources/learning-content/competence-datasource.js';
+import { competenceDatasource } from '../datasources/learning-content/competence-datasource.js';
 import { NotFoundError } from '../../../../lib/domain/errors.js';
 import { PIX_ORIGIN } from '../../../../lib/domain/constants.js';
 import { LOCALE } from '../../../../src/shared/domain/constants.js';
